added authCheck to routes/states
authCheck can be added as parameter to a route/state. It must be a function that is passed the current user and (in case of UI router the params of the state to be checked) and that returns true of access to this state is allowed for the passed user or false if access to this state should be denied to the passed user.